Remove llmc_py, single file
